requirements.txt
pillow==5.4.1
